fix(replay-clip): Ensure fullscreen analytics event is captured for replay clips
