+++ static/app/components/replays/replayFullscreenButton.tsx
@@ -1,6 +1,13 @@
+import {useCallback} from 'react';
+import screenfull from 'screenfull';
+
 import {Button} from 'sentry/components/button';
+import {useReplayContext} from 'sentry/components/replays/replayContext';
 import {IconContract, IconExpand} from 'sentry/icons';
 import {t} from 'sentry/locale';
+import {trackAnalytics} from 'sentry/utils/analytics';
+import useOrganization from 'sentry/utils/useOrganization';
+import {useUser} from 'sentry/utils/useUser';
 import useIsFullscreen from 'sentry/utils/window/useIsFullscreen';
 
 type Props = {
@@ -8,7 +15,28 @@ type Props = {
 };
 
 export function ReplayFullscreenButton({toggleFullscreen}: Props) {
+  const organization = useOrganization();
+  const user = useUser();
   const isFullscreen = useIsFullscreen();
+  const {analyticsContext} = useReplayContext();
+
+  const handleFullscreenToggle = useCallback(() => {
+    trackAnalytics('replay.toggle-fullscreen', {
+      organization,
+      context: analyticsContext,
+      user_email: user.email,
+      fullscreen: !isFullscreen,
+    });
+    toggleFullscreen();
+  }, [analyticsContext, isFullscreen, organization, toggleFullscreen, user.email]);
+
+  // If the browser supports going fullscreen or not. iPhone Safari won't do
+  // it. https://caniuse.com/fullscreen
+  const fullscreenSupported = screenfull.isEnabled;
+
+  if (!fullscreenSupported) {
+    return null;
+  }
 
   return (
     <Button
@@ -16,7 +44,7 @@ export function ReplayFullscreenButton({toggleFullscreen}: Props) {
       title={isFullscreen ? t('Exit full screen') : t('Enter full screen')}
       aria-label={isFullscreen ? t('Exit full screen') : t('Enter full screen')}
       icon={isFullscreen ? <IconContract size="sm" /> : <IconExpand size="sm" />}
-      onClick={toggleFullscreen}
+      onClick={handleFullscreenToggle}
     />
   );
 }
